backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-dotenv
jinja2
python-multipart
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
httpx
gspread

